+++ source/TimeSeries/IntegrationTests/Fixtures/AuthorizationConfiguration.cs
