+++ source/TimeSeries/IntegrationTests/Fixtures/ClientCredentialsSettings.cs
