+++ source/TimeSeries/IntegrationTests/Fixtures/TimeSeriesFunctionAppCollectionFixture.cs
