+++ source/TimeSeries/IntegrationTests/Fixtures/TimeSeriesFunctionAppFixture.cs
