+++ source/TimeSeries/IntegrationTests/IntegrationTests.csproj
